feat: add AppendBuffer implementation.
Includes support for doing writes, rollbacks and reads over the underlying temp file. The API of this struct has been design to follow the needs of the AppendService client code (doing writes) and the needs of the AsyncAppend internal code (rollbacks and reads), allowing a concurrent use of the same temp file from both the AppendService and AsyncAppend.

  It will need some extensions when developing the `AsyncAppend` code (like an explicit flush) but the foundations are there.